+++ src/main/java/application/Figures/Figure.java
@@ -9,8 +9,8 @@ import javafx.scene.control.Control;
 import javafx.scene.paint.Color;
 
 public abstract class Figure {
-    public abstract int[] getCoordinate();
-    public abstract void setCoordinate(int[] newCoordinates);
+    public abstract double[] getCoordinate();
+    public abstract void setCoordinate(double[] newCoordinates);
     public abstract void changePosition(double deltaX, double deltaY);
     public abstract Color getColor();
     public abstract void setColor(Color color);
