+++ src/main/java/application/Primitives/Primitive.java
@@ -5,8 +5,8 @@ import javafx.scene.Node;
 import javafx.scene.paint.Color;
 
 public abstract class Primitive {
-    public abstract int[] getCoordinate();
-    public abstract void setCoordinate(int[] newCoordinates);
+    public abstract double[] getCoordinate();
+    public abstract void setCoordinate(double[] newCoordinates);
     public abstract void changePosition(double deltaX, double deltaY);
     public abstract Color getColor();
     public abstract void setColor(Color color);
